Code clean-up prior to 4.1 release.

+++ webapp/src/main/java/edu/rpi/tw/escience/semanteco/impl/ModuleManagerImpl.java
@@ -6,7 +6,10 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.net.URI;
+import java.security.AccessController;
+import java.security.PrivilegedAction;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.Enumeration;
 import java.util.HashMap;
@@ -129,8 +132,7 @@ public class ModuleManagerImpl implements ModuleManager, FileListener {
 				module.visit(ui, request);
 			}
 			catch(Exception e) {
-				request.getLogger().warn("Module '"+module.getName()+"' threw an unexpected exception. Things may work incorrectly during the remainder of this request.", e);
-			
+				logException(request.getLogger(), module, e);
 			}
 		}
 	}
@@ -144,9 +146,8 @@ public class ModuleManagerImpl implements ModuleManager, FileListener {
 					module.visit((OntModel)model, request, domain);
 				}
 				catch(Exception e) {
-					request.getLogger().warn("Module '"+module.getName()+"' threw an unexpected exception. Things may work incorrectly during the remainder of this request.", e);
-					request.getLogger().warn("Stacktrace is: " + e.getStackTrace());
-
+					logException(request.getLogger(), module, e);
+					request.getLogger().warn("Stacktrace is: " + Arrays.toString(e.getStackTrace()));
 				}
 			}
 		}
@@ -161,8 +162,8 @@ public class ModuleManagerImpl implements ModuleManager, FileListener {
 					module.visit((Model)model, request, domain);
 				}
 				catch(Exception e) {
-					request.getLogger().warn("Module '"+module.getName()+"' threw an unexpected exception. Things may work incorrectly during the remainder of this request.", e);
-					request.getLogger().warn("Stacktrace is: " + e.getStackTrace());
+					logException(request.getLogger(), module, e);
+					request.getLogger().warn("Stacktrace is: " + Arrays.toString( e.getStackTrace() ));
 				}
 			}
 		}
@@ -194,7 +195,7 @@ public class ModuleManagerImpl implements ModuleManager, FileListener {
 					module.visit(query, request);
 				}
 				catch(Exception e) {
-					request.getLogger().warn("Module '"+module.getName()+"' threw an unexpected exception. Things may work incorrectly during the remainder of this request.", e);
+					logException(request.getLogger(), module, e);
 				}
 			}
 		}
@@ -207,17 +208,22 @@ public class ModuleManagerImpl implements ModuleManager, FileListener {
 	}
 
 	@Override
-	public final void fileCreated(FileChangeEvent event) {
+	public final void fileCreated(final FileChangeEvent event) {
 		log.trace("fileCreated");
 		log.debug("Observed file created: "+event.getFile().getName().getPath());
 		if(event.getFile().getName().getExtension().equals("jar")) {
 			ModuleClassLoader loader = null;
-			try {
-				loader = new ModuleClassLoader(event.getFile().getName().getPath());
-			}
-			catch(IllegalArgumentException e) {
-				return;
-			}
+			loader = AccessController.doPrivileged(new PrivilegedAction<ModuleClassLoader>() {
+				@Override
+				public ModuleClassLoader run() {
+					try {
+						return new ModuleClassLoader(event.getFile().getName().getPath());
+					}
+					catch(IllegalArgumentException e) {
+						return null;
+					}
+				}
+			});
 			if(loader != null) {
 				log.debug("Generated classloader");
 				classLoaders.put(event.getFile().getName().getPath(), loader);
@@ -229,9 +235,21 @@ public class ModuleManagerImpl implements ModuleManager, FileListener {
 			}
 		}
 	}
-	
+
+	protected final Module instantiateModule(Class<? extends Module> clazz) {
+		try {
+			log.debug("Generating new module instance");
+			return clazz.newInstance();
+		} catch (InstantiationException e) {
+			log.warn(MODULE_ERROR + clazz.getSimpleName(), e);
+		} catch (IllegalAccessException e) {
+			log.warn(MODULE_ERROR + clazz.getSimpleName(), e);
+		}
+		return null;
+	}
+
 	/**
-	 * INstantiates the module classes
+	 * Instantiates the module classes
 	 * @param loader
 	 * @param jarPath
 	 * @param resPath
@@ -240,15 +258,7 @@ public class ModuleManagerImpl implements ModuleManager, FileListener {
 			final String jarPath, final String resPath) {
 		Set<Class<? extends Module>> newModules = loader.getModules();
 		for(Class<? extends Module> i : newModules) {
-			Module module = null;
-			try {
-				log.debug("Generating new module instance");
-				module = i.newInstance();
-			} catch (InstantiationException e) {
-				log.warn(MODULE_ERROR+i.getSimpleName(), e);
-			} catch (IllegalAccessException e) {
-				log.warn(MODULE_ERROR+i.getSimpleName(), e);
-			}
+			Module module = instantiateModule(i);
 			if(module != null) {
 				InputStream properties = null;
 				JarFile jar = null;
@@ -259,9 +269,7 @@ public class ModuleManagerImpl implements ModuleManager, FileListener {
 						properties = jar.getInputStream(entry);
 					}
 				}
-				catch(IOException e) {
-					
-				}
+				catch(IOException e) { }
 				finally {
 					installModule(module, resPath, properties);
 					try {
@@ -284,15 +292,20 @@ public class ModuleManagerImpl implements ModuleManager, FileListener {
 			}
 		}
 	}
-	
+
+	protected final boolean ensureExists(final File dir) {
+		return dir.exists() || dir.mkdirs();
+	}
+
 	protected final String explodeJar(final String path) {
 		log.trace("Exploding jar");
 		JarFile jar = null;
 		try {
 			final String name = path.substring(path.lastIndexOf('/')+1, path.length()-".jar".length());
 			final File resDir = new File(this.path+"/../../resources/"+name+"/");
-			if(!resDir.exists()) {
-				resDir.mkdirs();
+			if(!resDir.exists() && !resDir.mkdirs()) {
+				log.error("Unable to create resource directory for module. Aborting...");
+				return null;
 			}
 			jar = new JarFile(path);
 			Enumeration<JarEntry> entries = jar.entries();
@@ -302,7 +315,10 @@ public class ModuleManagerImpl implements ModuleManager, FileListener {
 				if(entry.getName().startsWith(RES_DIR) &&
 						!entry.getName().endsWith("/")) {
 					File dest = new File(resDir, entry.getName().substring(RES_DIR.length()));
-					new File(dest.getParent()).mkdirs();
+					if(ensureExists(new File(dest.getParent()))) {
+						log.error("Unable to create subdirectory for resources. Aborting...");
+						return null;
+					}
 					copy(jar.getInputStream(entry), new FileOutputStream(dest));
 				}
 			}
@@ -463,7 +479,7 @@ public class ModuleManagerImpl implements ModuleManager, FileListener {
 		return new ArrayList<Domain>(knownDomains.values());
 	}
 	
-	protected void buildDomain() {
+	protected final void buildDomain() {
 		log.trace("buildDomain");
 		final Class<? extends ProvidesDomain> providerClass = ProvidesDomain.class;
 		moduleDomainMap = new HashMap<Module, List<Domain>>();
@@ -474,10 +490,10 @@ public class ModuleManagerImpl implements ModuleManager, FileListener {
 				ProvidesDomain provider = (ProvidesDomain)m;
 				List<Domain> domains = null;
 				try {
-					domains = provider.getDomains(new DummyRequest());
+					domains = provider.getDomains(new DummyRequest(this));
 				}
 				catch(Exception e) {
-					log.warn("Module '"+m.getName()+"' threw an unexpected exception. Things may work incorrectly during the remainder of this request.", e);
+					logException(log, m, e);
 				}
 				if(domains != null) {
 					moduleDomainMap.put(m, domains);
@@ -517,5 +533,11 @@ public class ModuleManagerImpl implements ModuleManager, FileListener {
 			}
 		}
 		return result;
+	}
+
+	protected final void logException(Logger log, Module module, Exception e) {
+		log.warn("Module '" + module.getName() +
+				"' threw an unexpected exception. Things may work incorrectly" +
+				" during the remainder of this request.", e);
 	}
 }
